abstract getting first item and setting item state out from Overview.jsx and into App.jsx, so that every widget can use its functionality
